server/requirements.txt
Django==4.2.3
psycopg2-binary
gunicorn
django-rest-framework
djangorestframework
python-decouple
django-cors-headers
djangorestframework-simplejwt